refactor: 상수 도메인으로 분리
