+++ src/constants/racingCarGame.js
@@ -1,0 +1,7 @@
+export const GAME_INFO_MESSAGE = {
+  START: '경주할 자동차 이름을 입력하세요.'
+}
+
+export const GAME_ERROR_MESSAGE = {
+  GAME_TERMINATE_OF_ERROR: '오류로 인해 게임이 종료되었습니다!'
+}
